More bugfix
git-svn-id: https://svn.win.tue.nl/repos/cpntools/AccessCPN@249 83e89a13-d7ff-4fff-b021-40a13f624d75

+++ trunk/org.cpntools.accesscpn.model.tests/src/org/cpntools/accesscpn/model/demos/statespacegenerator/EfficientStateSet.java
@@ -39,7 +39,7 @@ public class EfficientStateSet implements CompressedStateSet {
 			        final long l) throws StorageException, IOException {
 				final InputStream stream = store.getStreamForObject(l);
 				for (final byte b : object.getBytes()) {
-					if (stream.read() != b) { return false; }
+					if ((byte) stream.read() != b) { return false; }
 				}
 				return true;
 			}
